+++ 42/push_swap/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: sgoldenb <sgoldenb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 22:40:57 by sgoldenb          #+#    #+#             */
-/*   Updated: 2023/08/31 19:49:12 by sgoldenb         ###   ########.fr       */
+/*   Updated: 2023/09/01 23:26:06 by sgoldenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ char				*ft_strnstr(const char *big,
 int					ft_strcmp(const char *s1, const char *s2);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 
-int					ft_atoi(const char *str);
+long int			ft_atoi(const char *str);
 
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
